+++ qiskit/primitives/estimator.py
@@ -22,11 +22,12 @@ import numpy as np
 
 from qiskit.circuit import Parameter, QuantumCircuit
 from qiskit.exceptions import QiskitError
+from qiskit.opflow import PauliSumOp
 from qiskit.primitives import BaseEstimator, EstimatorResult
+from qiskit.quantum_info import Statevector
 from qiskit.quantum_info.operators.base_operator import BaseOperator
-from qiskit.quantum_info.states import Statevector
 
-from .utils import PauliSumOp, init_circuit, init_observable
+from .utils import init_circuit, init_observable
 
 
 class Estimator(BaseEstimator):
