+++ qiskit/primitives/sampler.py
@@ -21,7 +21,7 @@ from typing import Optional
 from qiskit.circuit import Parameter, QuantumCircuit
 from qiskit.exceptions import QiskitError
 from qiskit.primitives import BaseSampler, SamplerResult
-from qiskit.quantum_info.states import Statevector
+from qiskit.quantum_info import Statevector
 from qiskit.result import QuasiDistribution
 
 from .utils import init_circuit, final_measurement_mapping
