+++ qiskit/primitives/utils.py
@@ -15,44 +15,11 @@ Utility functions for primitives
 
 from __future__ import annotations
 
-import sys
-
 from qiskit.circuit import ParameterExpression, QuantumCircuit
 from qiskit.extensions.quantum_initializer.initializer import Initialize
-from qiskit.quantum_info.operators import SparsePauliOp
+from qiskit.opflow import PauliSumOp
+from qiskit.quantum_info import SparsePauliOp, Statevector
 from qiskit.quantum_info.operators.base_operator import BaseOperator
-from qiskit.quantum_info.states import Statevector
-
-if sys.version_info >= (3, 8):
-    # pylint: disable=no-name-in-module
-    from typing import Protocol, runtime_checkable
-else:
-    from typing_extensions import Protocol, runtime_checkable
-
-
-@runtime_checkable
-class PauliSumOp(Protocol):
-    """Protocol for opflow.PauliSumOp."""
-
-    @property
-    def primitive(self) -> SparsePauliOp:
-        """
-        The SparsePauliOp which defines the behavior of the underlying function.
-
-        Returns:
-             The primitive SparsePauliOp.
-        """
-        ...
-
-    @property
-    def coeff(self) -> complex | ParameterExpression:
-        """
-        The scalar coefficient multiplying the Operator.
-
-        Returns:
-              The coefficient.
-        """
-        ...
 
 
 def init_circuit(state: QuantumCircuit | Statevector) -> QuantumCircuit:
